Move ColorPatterns into DisplayPatternLib
