--- lib/ColorPatternLib/ColorPatternLib.h
@@ -1,14 +1,0 @@
-#ifndef COLORPATTERNLIB_H
-#define COLORPATTERNLIB_H
-
-// This header file is only in place to link to the other files in the common lib.
-
-#include "ColorPattern.h"
-#include "ColorFadePattern.h"
-#include "RainbowColorPattern.h"
-#include "SingleColorPattern.h"
-#include "TwoColorFadePattern.h"
-#include "TwoColorPattern.h"
-#include "BackgroundPlusThree.h"
-
-#endif
