+++ lib/DisplayPatternLib/src/ColorPatterns/BackgroundPlusThree.cpp
