+++ lib/DisplayPatternLib/src/ColorPatterns/BackgroundPlusThree.h
