+++ lib/DisplayPatternLib/src/ColorPatterns/ColorFadePattern.cpp
