+++ lib/DisplayPatternLib/src/ColorPatterns/ColorFadePattern.h
