+++ lib/DisplayPatternLib/src/ColorPatterns/ColorPattern.h
