+++ lib/DisplayPatternLib/src/ColorPatterns/RainbowColorPattern.cpp
