+++ lib/DisplayPatternLib/src/ColorPatterns/RainbowColorPattern.h
