+++ lib/DisplayPatternLib/src/ColorPatterns/SingleColorPattern.cpp
