+++ lib/DisplayPatternLib/src/ColorPatterns/SingleColorPattern.h
