+++ lib/DisplayPatternLib/src/ColorPatterns/TwoColorFadePattern.cpp
