+++ lib/DisplayPatternLib/src/ColorPatterns/TwoColorFadePattern.h
