+++ lib/DisplayPatternLib/src/ColorPatterns/TwoColorPattern.cpp
