+++ lib/DisplayPatternLib/src/ColorPatterns/TwoColorPattern.h
