+++ lib/DisplayPatternLib/src/PatternFactory.h
@@ -8,7 +8,13 @@
 #include "DisplayPatterns\RandomDisplayPattern.h"
 #include "DisplayPatterns\CenterOutDisplayPattern.h"
 #include "DisplayPatterns\LowPowerDisplayPattern.h"
-#include <ColorPatternLib.h>
+#include "ColorPatterns\ColorPattern.h"
+#include "ColorPatterns\ColorFadePattern.h"
+#include "ColorPatterns\RainbowColorPattern.h"
+#include "ColorPatterns\SingleColorPattern.h"
+#include "ColorPatterns\TwoColorFadePattern.h"
+#include "ColorPatterns\TwoColorPattern.h"
+#include "ColorPatterns\BackgroundPlusThree.h"
 #include <PixelBuffer.h>
 #include <PatternData.h>
 #include <vector>
